1208_2 cluster

+++ ft_printf/source/specifier/spcf_c.c
@@ -6,14 +6,14 @@
 /*   By: donghyu2 <donghyu2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 00:53:56 by donghyu2          #+#    #+#             */
-/*   Updated: 2022/12/02 01:41:25 by donghyu2         ###   ########.fr       */
+/*   Updated: 2022/12/08 18:09:55 by donghyu2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 #include "libft.h"
 
-size_t	spcf_c(va_list *ptr)
+int	spcf_c(va_list *ptr)
 {
 	return (ft_putchar_len((unsigned char)va_arg(*ptr, int)));
 }
