+++ ft_printf/source/specifier/spcf_d.c
@@ -6,14 +6,14 @@
 /*   By: donghyu2 <donghyu2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 00:54:00 by donghyu2          #+#    #+#             */
-/*   Updated: 2022/12/02 01:42:42 by donghyu2         ###   ########.fr       */
+/*   Updated: 2022/12/08 18:10:00 by donghyu2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 #include "libft.h"
 
-size_t	spcf_d(va_list *ptr)
+int	spcf_d(va_list *ptr)
 {
 	return (ft_putnbr_len(va_arg(*ptr, int), 0));
 }
