+++ ft_printf/source/specifier/spcf_i.c
@@ -6,14 +6,14 @@
 /*   By: donghyu2 <donghyu2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 00:54:03 by donghyu2          #+#    #+#             */
-/*   Updated: 2022/12/02 01:42:35 by donghyu2         ###   ########.fr       */
+/*   Updated: 2022/12/08 18:11:18 by donghyu2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 #include "libft.h"
 
-size_t	spcf_i(va_list *ptr)
+int	spcf_i(va_list *ptr)
 {
 	return (ft_putnbr_len(va_arg(*ptr, int), 0));
 }
