+++ ft_printf/source/specifier/spcf_p.c
@@ -6,14 +6,14 @@
 /*   By: donghyu2 <donghyu2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 00:54:11 by donghyu2          #+#    #+#             */
-/*   Updated: 2022/12/06 15:29:25 by donghyu2         ###   ########.fr       */
+/*   Updated: 2022/12/08 18:11:11 by donghyu2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 #include "libft.h"
 
-size_t	spcf_p(va_list *ptr)
+int	spcf_p(va_list *ptr)
 {
 	unsigned long	addr;
