+++ ft_printf/source/specifier/spcf_s.c
@@ -6,14 +6,14 @@
 /*   By: donghyu2 <donghyu2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 00:54:28 by donghyu2          #+#    #+#             */
-/*   Updated: 2022/12/02 01:42:11 by donghyu2         ###   ########.fr       */
+/*   Updated: 2022/12/08 18:11:06 by donghyu2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 #include "libft.h"
 
-size_t	spcf_s(va_list *ptr)
+int	spcf_s(va_list *ptr)
 {
 	char	*str;
