+++ ft_printf/source/specifier/spcf_u.c
@@ -6,16 +6,16 @@
 /*   By: donghyu2 <donghyu2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 00:54:26 by donghyu2          #+#    #+#             */
-/*   Updated: 2022/12/01 22:30:54 by donghyu2         ###   ########.fr       */
+/*   Updated: 2022/12/08 18:10:46 by donghyu2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 #include "libftprintf.h"
 
-static size_t	ft_putnbr_uint_len(unsigned int num, size_t len);
+static int	ft_putnbr_uint_len(unsigned int num, int len);
 
-size_t	spcf_u(va_list *ptr)
+int	spcf_u(va_list *ptr)
 {
 	unsigned int	num;
 
@@ -23,7 +23,7 @@ size_t	spcf_u(va_list *ptr)
 	return (ft_putnbr_uint_len(num, 0));
 }
 
-static size_t	ft_putnbr_uint_len(unsigned int num, size_t len)
+static int	ft_putnbr_uint_len(unsigned int num, int len)
 {
 	if (num < 10)
 		len += write(1, &"0123456789"[num], 1);
