+++ ft_printf/source/specifier/spcf_x_upper.c
@@ -6,14 +6,14 @@
 /*   By: donghyu2 <donghyu2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 00:54:21 by donghyu2          #+#    #+#             */
-/*   Updated: 2022/12/02 01:41:40 by donghyu2         ###   ########.fr       */
+/*   Updated: 2022/12/08 18:10:17 by donghyu2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 #include "libft.h"
 
-size_t	spcf_x_upper(va_list *ptr)
+int	spcf_x_upper(va_list *ptr)
 {
 	unsigned int	num;
